solve recalculations and equation deletion
i observe recalculations are caused by excel when nested functions are used with the expand set to true as it writes to excel at each nested result as it should just only writes at the end of the calculations. At the same time this solves the problem of matrix being resized at every function.